Fix the kind function.
It was broken in that it relied on the order of the groups. Added some
logic to filter by list length and whatnot. Listen, I'm tired. Shut up.

Add tests for fullHouse and various kinds which exercise this. fullHouse
in particular was a great test case for matching against three of a kind
AND two of a kind. It's passing now. I think.