First attempt at implementing shuffling.
Turns out that Data.Random.Extras is overkill for what I'm doing, but I
thought I'd keep this here as a first attempt.

The trap I'm in now is that the result of shuffling is inside an RVar
and RVars seem… problematic when it comes to getting things *out* of
them.

A fellow on haskell-cafe has suggested there's a more straightforward
way to do this. Possibly generating a random number based on length of
deck, taking the item at that index out, and finishing would be the
best way.