Nix State, RVar.
Thanks to a fellow on haskell-cafe, I've got a new implementation of
dealHands as well as shuffling.

The dealing is quite clever:
- create an infinite list of cards, each with n cards fewer than the
last;
- split off m from the list of cards, one for each player/hand;
- for each of the list of cards, take n cards;
- for the infinite list of remaining cards, take the head

The last two parts use an Arrow to work with the tuple, composing a
single function from two functions, one to operate on the fst item in
tuple, the other to operate on snd.

Upcoming is a refactor to make Cards into their own datatype.